NavigationItemのenumを別ファイルに
